+++ common/src/main/java/net/apartium/cocoabeans/Ensures.java
@@ -10,6 +10,7 @@
 
 package net.apartium.cocoabeans;
 
+import org.jetbrains.annotations.ApiStatus;
 import java.util.Collection;
 import java.util.Map;
 
@@ -200,7 +201,6 @@ public class Ensures {
     }
 
     /**
-     *
      * Makes sure given condition is true
      *
      * @param bool the condition
@@ -211,8 +211,46 @@ public class Ensures {
         if (!bool)
             throw ex == null ? new RuntimeException() : ex;
 
-
     }
 
+    /**
+     * Makes sure given condition is false
+     *
+     * @param bool the condition
+     * @param ex   the exception to throw if the boolean is true
+     * @throws RuntimeException ex if argument does not match requested precondition
+     *
+     */
+    @ApiStatus.AvailableSince("0.0.39")
+    public static void isFalse(boolean bool, RuntimeException ex) {
+        if (bool)
+            throw ex == null ? new RuntimeException() : ex;
+    }
+
+
+    /**
+     * Makes sure given condition is false
+     *
+     * @param bool the condition
+     * @throws RuntimeException if argument does not match requested precondition
+     */
+    @ApiStatus.AvailableSince("0.0.39")
+    public static void isFalse(boolean bool) {
+        isFalse(bool, new RuntimeException());
+    }
+
+    /**
+     * Makes sure given condition is false
+     *
+     * @param bool the condition
+     * @param message the message to send with the exception, "+-" will be replaced with "must be false"
+     * @throws RuntimeException if argument does not match requested precondition
+     *
+     */
+    @ApiStatus.AvailableSince("0.0.39")
+    public static void isFalse(boolean bool, String message) {
+        message = message.replaceAll("\\+-", "must be false");
+        isFalse(bool, new RuntimeException(message));
+    }
 
 }