+++ spigot/src/test/java/net/apartium/cocoabeans/spigot/craftserver/SpigotCraftServer_v1_8_R3.java
@@ -1,0 +1,11 @@
+package net.apartium.cocoabeans.spigot.craftserver;
+
+import net.apartium.cocoabeans.spigot.NMSUtils;
+
+/**
+ * A stubbing class for checking the functionality of NMSUtils#containsVersion(String)
+ * @see NMSUtils
+ */
+public class SpigotCraftServer_v1_8_R3 extends SpigotCraftServer {
+
+}